+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * API change: Renamed `label` to `valueLabel` and `description` to `valueDescription` in `AnnotatedValue`.
 * Added `AnnotatedValue.typeLabel` and `AnnotatedValue.typeDescription`.
 * Added `ProvidedMeta.optional`.
+* Removed support for features in an action arguments provision result itself (`ProvidedValue`). Such features can be provided in an `AnnotatedValue`.
+* Removed experimental support for pageable value set and element value set.
 
 ## 1.13.0 (2019-10-30)